Move FetchProperties interface to 3_0 only
