--- src/dbus_api/filesystem/fetch_properties_2_0/mod.rs
@@ -1,4 +1,0 @@
-mod api;
-mod methods;
-
-pub use api::{get_all_properties_method, get_properties_method};
